Remove generated portion of package.json
Some parts of the `package.json` are generated when installing via
`npm`.  These include the `readme` as well as other data about the
project.  We can safely remove these sections since they do not need to
be in the `package.json` file, and are only added there when the module
is installed via `npm`.